fix: clear visualization pr merge checks
